Main Activity UI Update
Added drawable gradient_bottom_to_top to have a dark gray pallete that fades to transparent from the bottom to the top. Will use it to put on TextView background that have photos behind text.
Updated the color palette and implemented it in the main style theme, added a lighter tone green, and an dark orange for its accent.
Tablet now has 2 rows to avoid empty space on the bottom
Updated UI in list_item_article for the picture to take all the card's space, keep its ration, and have the important texts on top of the preview photo.
Created a toolbar layout "toolbar_article_list" to avoid -v21 warnings when was using elevation, and included the new toolbar in "activity_article_list"
Added 2 style attribute for Header and Body text types for the article_items layout
Added dimens sizes for Header and Body text types
Added ConstraintLayout androidx library

This changes to the UI gives the main activity a more Material look.